Code review feedback

+++ include/message.hpp
@@ -60,11 +60,11 @@ struct Message
     /* Session ID and accounting record number are always filled in by
        the session manager. */
     std::string session_id;
-    long accounting_record_number;
+    uint32_t accounting_record_number;
     std::string timer_id;
 
-    int interim_interval;
-    int session_refresh_time;
+    uint32_t interim_interval;
+    uint32_t session_refresh_time;
 };
 
 #endif
